Update App.js
On submitting the form, the task is added to the todo array. The code uses the map to iterate through the todo array, and renders each task as a list item. Using useState, this component registers a state — value — and a function for updating it — setTodo. The handleSubmit handler will prevent the default action that would normally be taken on the form and add a new Task using the latest value that is in the input field.